move deletecell to own component
