+++ src/components/Columns.tsx
@@ -4,6 +4,7 @@ import { TodosData } from 'models/todos'
 import PriorityCell from './PriorityCell'
 import DueDateCell from './DueDateCell'
 import CategoryCell from './CategoryCell'
+import DeleteCell from './DeleteCell'
 import { Button } from './ui/button'
 import useDeleteTodo from '@/hooks/use-delete-todo'
 
@@ -34,15 +35,6 @@ export const columns: ColumnDef<TodosData>[] = [
   {
     accessorKey: 'delete',
     header: 'Delete',
-    cell: ({ row }) => {
-      const id = row.original.id
-      const deleteTodo = useDeleteTodo()
-
-      const deleteClickHandler = () => {
-        deleteTodo.mutate(id)
-      }
-
-      return <Button onClick={deleteClickHandler}>Delete</Button>
-    },
+    cell: DeleteCell,
   },
 ]
